Rename stack/helpers/matches.rs; fix clippy
