+++ synthesizer/process/src/stack/helpers/mod.rs
@@ -16,6 +16,6 @@
 use super::*;
 
 mod initialize;
-mod matches;
 mod sample;
+mod stack_trait;
 mod synthesize;
