+++ synthesizer/process/src/stack/helpers/stack_trait.rs
